+++ app/javascript/authentication/components/IdleTimeout.tsx
