+++ app/javascript/layouts/withAppSetup.tsx
@@ -1,6 +1,6 @@
 import React from "react"
 
-import IdleTimeout from "../authentication/components/IdeTimeout"
+import IdleTimeout from "../authentication/components/IdleTimeout"
 import UserProvider from "../authentication/context/UserProvider"
 
 // Ignore linting error on 'object' type, because we can't use Record<string, unknown> here.
